Added comments to structs / methods

+++ group.go
@@ -1,12 +1,23 @@
 package main
 
+import "errors"
+
+// A Group represents a collection of Leaders.
+// Groups should only be accessible to the Group's owner.
 type Group struct {
-	GroupID int
-	OwnerID int
-	Name    string
-	Members []*Leader
+	GroupID int       // Unique id of the Group
+	OwnerID int       // ID of the Group's owner
+	Name    string    // Name of the Group
+	Members []*Leader // Members of the Group added by the owner.
 }
 
-func (g *Group) Delete() {
+// Deletes the Group from the database.
+// Returns an error if the leader referenced is not the Group's owner
+func (g *Group) Delete(leaderID int) error {
+	if leaderID != g.OwnerID {
+		return errors.New("the leaderID provided does not match the Group's OwnerID")
+	}
+
 	DeleteGroupDB(g.GroupID)
+	return nil
 }
